// compumundo-app-master/src/app/interfaces/ruta.ts
export interface Ruta {
    estado?: number;
    IdCliente?: string;
    NombreCliente?: string;
    id?: string;
    usuario?: string;
    "CODCliente"?: string
    "Nit"?: string
    "Nombre"?: string
    "RazonSocial"?: string
    "Direccion"?: string
    "Telefono"?: string
    "Observaciones"?: string
    "IDTipoPrecio"?: number
    "visitada"?: boolean
    DiasVencimiento?: number
    SerieFactura?: string;
    IdFactura?: string;
    Serie?: string;
    IDDocumento?: number;
    recibo?: boolean;
    SaldoPagado?: number;
    Ruta?: string;
}